Remove unsubscribed or closed channels
This seems only an issue for IM channels. However there isn't a
`excludeArchived` for IM channels. So we will only add users that
aren't deleted to the UserCache. When we are adding the IM channels
we will check if the name is present in the UserCache, when it isn't
we will ignore it. Resulting in channels without IM channels that
are from delete users.

Fixes #3