Add desktop notification support
This commit adds optional support for desktop notifications via the
`notificator` Go library.

To control desktop notifications, an optional field `notify` is added to
`slack-term.json` config file. Setting it to true will enable sending
desktop notifications for all received messages.

To achieve this, a function `NotifyMessages` was added to the Slack
service, which accepts a channel ID and a list of messages, and sends a
desktop notification containing the full message text to the user via
the `notificator` library.containing the full message text to the user
via the `notificator` library. Aditionally, the function
`actionNewMessage` in the event handler was amended to call
`NotifyMessages`, and tweaked to accept a list of messages alongside a
channel ID, so that information can be passed to `NotifyMessages`.

The `notificator` library is multi-platform, supporting Linux, Windows
and OS X. The OS notification services will truncate the message to an
appropriate length in the UI.